bc7enc: Optimize "find approximate selector" branch chains
Several BC7 code paths have branch chains which sequentially
compare a value against an array of thresholds. These chains
are long enough that compilers have trouble converting them to
branchless operations.

In all of these code paths, the value produced by the branch chain
is a direct dependency of of the subsequent code. This often results
in a pipeline stall, because the branches can't be easily predicted.

To improve this, convert each branch chain to a branchless loop.
Compiler optimizations will inline and unroll the loop, significantly
improving the codgen in these code paths and making room for further
compiler optimizations (such as auto-vectorization).